examples/epsilon: Restructure packages
