+++ example/epsilon/bookstore/app/authors/books/books.go
@@ -1,4 +1,4 @@
-package authors
+package books
 
 import "github.com/TheFellow/go-arch-lint/example/epsilon/bookstore/app/books"
